+++ packages/block-library/src/group/save.js
@@ -3,12 +3,6 @@
  */
 import { useInnerBlocksProps, useBlockProps } from '@wordpress/block-editor';
 
-export default function save( { attributes } ) {
-	const { tagName: Tag, role } = attributes;
-	return (
-		<Tag
-			{ ...useInnerBlocksProps.save( useBlockProps.save() ) }
-			role={ role }
-		/>
-	);
+export default function save( { attributes: { tagName: Tag } } ) {
+	return <Tag { ...useInnerBlocksProps.save( useBlockProps.save() ) } />;
 }